Isabella/SF2/requirements.txt
stable-baselines3==1.8.0
gymnasium==0.29.1
gym-retro==0.8.1
numpy==1.24.3
torch==2.0.1
traceback==0.1.0
stable-baselines3[extra]==1.8.0
